+++ code/backend/src/main/java/com/cs673/CareerForge/CareerForgeApplication.java
@@ -1,13 +1,13 @@
-package com.cs673.CurseForge;
+package com.cs673.CareerForge;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
 @SpringBootApplication(exclude = {SecurityAutoConfiguration.class })
-public class CurseForgeApplication {
+public class CareerForgeApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(CurseForgeApplication.class, args);
+		SpringApplication.run(CareerForgeApplication.class, args);
 	}
 }
