+++ code/backend/src/main/resources/application.properties
@@ -1,2 +1,2 @@
-spring.application.name=CurseForge
+spring.application.name=CareerForge
 server.port=8080
