+++ code/backend/src/test/java/com/cs673/CurseForge/CareerForgeApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class CurseForgeApplicationTests {
+class CareerForgeApplicationTests {
 
 	@Test
 	void contextLoads() {
